feat(frontend): roll out TanStack Query/Form split across remaining pages
