+++ apps/frontend/app/(admin)/admin/editions/[id]/participations/page.tsx
@@ -5,23 +5,13 @@ import { ConfirmDialog } from '@/components/common/ConfirmDialog';
 import { DataTable } from '@/components/common/DataTable';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import { ApiError, apiClient, throwIfError } from '@/lib/api/client';
-import { queryKeys } from '@/lib/query/keys';
-import { getApiErrorMessage } from '@/lib/utils/errors';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import {
+  type Participation,
+  useAdminParticipationsPage,
+} from '@/features/admin/participations/hooks';
 import type { ColumnDef } from '@tanstack/react-table';
 import { CheckIcon, PencilIcon, PlusIcon, Trash2Icon } from 'lucide-react';
-import { use, useState } from 'react';
-import { toast } from 'sonner';
-
-type Participation = {
-  id: string;
-  editionId: string;
-  universityId: string;
-  universityName: string;
-  teamName: string | null;
-  createdAt: unknown;
-};
+import { use } from 'react';
 
 export default function AdminParticipationsPage({
   params,
@@ -29,68 +19,21 @@ export default function AdminParticipationsPage({
   params: Promise<{ id: string }>;
 }) {
   const { id: editionId } = use(params);
-  const queryClient = useQueryClient();
-  const [selectedUniversityId, setSelectedUniversityId] = useState('');
-  const [teamName, setTeamName] = useState('');
-  const [editingId, setEditingId] = useState<string | null>(null);
-  const [editingTeamName, setEditingTeamName] = useState('');
-
-  const { data, isLoading } = useQuery({
-    queryKey: queryKeys.admin.participations(editionId, {}),
-    queryFn: async () => {
-      const result = await apiClient.GET('/api/admin/editions/{id}/participations', {
-        params: { path: { id: editionId } },
-      });
-      return throwIfError(result);
-    },
-  });
-
-  const createMutation = useMutation({
-    mutationFn: async () => {
-      const result = await apiClient.POST('/api/admin/editions/{id}/participations', {
-        params: { path: { id: editionId } },
-        body: { universityId: selectedUniversityId, teamName: teamName || undefined },
-      });
-      return throwIfError(result);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: queryKeys.admin.participations(editionId, {}) });
-      setSelectedUniversityId('');
-      setTeamName('');
-      toast.success('出場登録しました');
-    },
-    onError: (err) => toast.error(getApiErrorMessage(err)),
-  });
-
-  const updateMutation = useMutation({
-    mutationFn: async ({ id, teamName }: { id: string; teamName: string | null }) => {
-      const result = await apiClient.PUT('/api/admin/participations/{id}', {
-        params: { path: { id } },
-        body: { teamName: teamName || null },
-      });
-      return throwIfError(result);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: queryKeys.admin.participations(editionId, {}) });
-      setEditingId(null);
-      toast.success('更新しました');
-    },
-    onError: (err) => toast.error(getApiErrorMessage(err)),
-  });
-
-  const deleteMutation = useMutation({
-    mutationFn: async (id: string) => {
-      const result = await apiClient.DELETE('/api/admin/participations/{id}', {
-        params: { path: { id } },
-      });
-      if (!result.response.ok) throw new ApiError(result.response.status, result.error);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: queryKeys.admin.participations(editionId, {}) });
-      toast.success('削除しました');
-    },
-    onError: (err) => toast.error(getApiErrorMessage(err)),
-  });
+  const {
+    selectedUniversityId,
+    setSelectedUniversityId,
+    teamName,
+    setTeamName,
+    editingId,
+    setEditingId,
+    editingTeamName,
+    setEditingTeamName,
+    data,
+    isLoading,
+    createMutation,
+    updateMutation,
+    deleteMutation,
+  } = useAdminParticipationsPage(editionId);
   const columns: ColumnDef<Participation>[] = [
     { header: '大学名', accessorKey: 'universityName' },
     {
