+++ apps/frontend/app/(admin)/admin/universities/page.tsx
@@ -10,51 +10,17 @@ import {
   DialogTrigger,
 } from '@/components/ui/dialog';
 import { Input } from '@/components/ui/input';
-import { apiClient, throwIfError } from '@/lib/api/client';
-import { queryKeys } from '@/lib/query/keys';
-import { getApiErrorMessage } from '@/lib/utils/errors';
-import { useForm } from '@tanstack/react-form';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import {
+  type University,
+  useAdminUniversitiesList,
+  useCreateUniversityForm,
+} from '@/features/admin/universities/hooks';
 import type { ColumnDef } from '@tanstack/react-table';
 import { PlusIcon } from 'lucide-react';
 import { useState } from 'react';
-import { toast } from 'sonner';
-import { z } from 'zod';
-
-type University = {
-  id: string;
-  name: string;
-  slug: string;
-  createdAt: unknown;
-};
 
 function CreateUniversityDialog({ onClose }: { onClose: () => void }) {
-  const queryClient = useQueryClient();
-  const form = useForm({
-    defaultValues: { name: '', slug: '', ownerEmail: '' },
-    onSubmit: async ({ value }) => {
-      await mutation.mutateAsync(value);
-    },
-  });
-
-  const mutation = useMutation({
-    mutationFn: async (values: { name: string; slug: string; ownerEmail: string }) => {
-      const r = await apiClient.POST('/api/admin/universities', {
-        body: {
-          name: values.name,
-          slug: values.slug,
-          ownerEmail: values.ownerEmail || undefined,
-        },
-      });
-      return throwIfError(r);
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: queryKeys.admin.universities({}) });
-      toast.success('大学を作成しました');
-      onClose();
-    },
-    onError: (err) => toast.error(getApiErrorMessage(err)),
-  });
+  const { form, mutation, validators } = useCreateUniversityForm(onClose);
 
   return (
     <DialogContent>
@@ -68,10 +34,7 @@ function CreateUniversityDialog({ onClose }: { onClose: () => void }) {
         }}
         className='space-y-4'
       >
-        <form.Field
-          name='name'
-          validators={{ onChange: z.string().min(1, '名称を入力してください') }}
-        >
+        <form.Field name='name' validators={{ onChange: validators.name }}>
           {(field) => (
             <div className='space-y-1'>
               <label htmlFor={field.name} className='text-sm font-medium'>
@@ -90,15 +53,7 @@ function CreateUniversityDialog({ onClose }: { onClose: () => void }) {
             </div>
           )}
         </form.Field>
-        <form.Field
-          name='slug'
-          validators={{
-            onChange: z
-              .string()
-              .min(1, 'スラッグを入力してください')
-              .regex(/^[a-z0-9-]+$/, '半角英数字とハイフンのみ使用できます'),
-          }}
-        >
+        <form.Field name='slug' validators={{ onChange: validators.slug }}>
           {(field) => (
             <div className='space-y-1'>
               <label htmlFor={field.name} className='text-sm font-medium'>
@@ -117,12 +72,7 @@ function CreateUniversityDialog({ onClose }: { onClose: () => void }) {
             </div>
           )}
         </form.Field>
-        <form.Field
-          name='ownerEmail'
-          validators={{
-            onChange: z.string().email('有効なメールアドレスを入力してください').or(z.literal('')),
-          }}
-        >
+        <form.Field name='ownerEmail' validators={{ onChange: validators.ownerEmail }}>
           {(field) => (
             <div className='space-y-1'>
               <label htmlFor={field.name} className='text-sm font-medium'>
@@ -157,16 +107,7 @@ function CreateUniversityDialog({ onClose }: { onClose: () => void }) {
 
 export default function AdminUniversitiesPage() {
   const [dialogOpen, setDialogOpen] = useState(false);
-
-  const { data, isLoading } = useQuery({
-    queryKey: queryKeys.admin.universities({}),
-    queryFn: async () => {
-      const result = await apiClient.GET('/api/admin/universities', {
-        params: { query: { pageSize: 100 } },
-      });
-      return throwIfError(result);
-    },
-  });
+  const { data, isLoading } = useAdminUniversitiesList();
 
   const columns: ColumnDef<University>[] = [
     { header: '名称', accessorKey: 'name' },
