+++ apps/frontend/app/(authenticated)/account/settings/page.tsx
@@ -3,62 +3,17 @@
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
-import { useAuth, useInvalidateMe } from '@/contexts/AuthContext';
-import { authClient } from '@/lib/auth/client';
-import { useForm } from '@tanstack/react-form';
-import { useMutation } from '@tanstack/react-query';
-import { toast } from 'sonner';
-import { z } from 'zod';
+import { useAccountSettingsForms } from '@/features/account/settings/hooks';
 
 export default function AccountSettingsPage() {
-  const { user } = useAuth();
-  const invalidateMe = useInvalidateMe();
-
-  const profileForm = useForm({
-    defaultValues: { name: user?.name ?? '' },
-    onSubmit: async ({ value }) => {
-      await updateProfileMutation.mutateAsync(value);
-    },
-  });
-
-  const passwordForm = useForm({
-    defaultValues: { currentPassword: '', newPassword: '', confirmPassword: '' },
-    onSubmit: async ({ value }) => {
-      await changePasswordMutation.mutateAsync(value);
-    },
-  });
-
-  const updateProfileMutation = useMutation({
-    mutationFn: async (values: { name: string }) => {
-      const result = await authClient.updateUser({ name: values.name });
-      if (result.error) throw new Error(result.error.message);
-    },
-    onSuccess: async () => {
-      await invalidateMe();
-      toast.success('プロフィールを更新しました');
-    },
-    onError: (err) => toast.error(err instanceof Error ? err.message : '更新に失敗しました'),
-  });
-
-  const changePasswordMutation = useMutation({
-    mutationFn: async (values: {
-      currentPassword: string;
-      newPassword: string;
-      confirmPassword: string;
-    }) => {
-      const result = await authClient.changePassword({
-        currentPassword: values.currentPassword,
-        newPassword: values.newPassword,
-      });
-      if (result.error) throw new Error(result.error.message);
-    },
-    onSuccess: () => {
-      passwordForm.reset();
-      toast.success('パスワードを変更しました');
-    },
-    onError: (err) =>
-      toast.error(err instanceof Error ? err.message : 'パスワード変更に失敗しました'),
-  });
+  const {
+    user,
+    profileForm,
+    passwordForm,
+    updateProfileMutation,
+    changePasswordMutation,
+    validators,
+  } = useAccountSettingsForms();
 
   return (
     <div className='space-y-8 max-w-xl'>
@@ -77,10 +32,7 @@ export default function AccountSettingsPage() {
             className='space-y-4'
           >
             <div className='text-sm text-muted-foreground'>メールアドレス: {user?.email}</div>
-            <profileForm.Field
-              name='name'
-              validators={{ onChange: z.string().min(1, '名前を入力してください') }}
-            >
+            <profileForm.Field name='name' validators={{ onChange: validators.profileName }}>
               {(field) => (
                 <div className='space-y-1'>
                   <label htmlFor={field.name} className='text-sm font-medium'>
@@ -119,7 +71,7 @@ export default function AccountSettingsPage() {
           >
             <passwordForm.Field
               name='currentPassword'
-              validators={{ onChange: z.string().min(1, '現在のパスワードを入力してください') }}
+              validators={{ onChange: validators.currentPassword }}
             >
               {(field) => (
                 <div className='space-y-1'>
@@ -141,9 +93,7 @@ export default function AccountSettingsPage() {
             </passwordForm.Field>
             <passwordForm.Field
               name='newPassword'
-              validators={{
-                onChange: z.string().min(8, '新しいパスワードは8文字以上で入力してください'),
-              }}
+              validators={{ onChange: validators.newPassword }}
             >
               {(field) => (
                 <div className='space-y-1'>
@@ -168,10 +118,10 @@ export default function AccountSettingsPage() {
               validators={{
                 onChangeListenTo: ['newPassword'],
                 onChange: ({ value, fieldApi }) => {
-                  if (value !== fieldApi.form.getFieldValue('newPassword')) {
-                    return { message: 'パスワードが一致しません' };
-                  }
-                  return undefined;
+                  return validators.confirmPassword({
+                    value,
+                    newPassword: fieldApi.form.getFieldValue('newPassword'),
+                  });
                 },
               }}
             >
