+++ apps/frontend/app/(authenticated)/dashboard/page.tsx
@@ -5,51 +5,11 @@ import { StatusBadge } from '@/components/common/StatusBadge';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Skeleton } from '@/components/ui/skeleton';
-import { useOrganization } from '@/contexts/OrganizationContext';
-import { apiClient, throwIfError } from '@/lib/api/client';
-import { queryKeys } from '@/lib/query/keys';
-import { useQueries, useQuery } from '@tanstack/react-query';
+import { useDashboardData } from '@/features/dashboard/query';
 import Link from 'next/link';
 
 export default function DashboardPage() {
-  const { organizationId, currentOrg } = useOrganization();
-
-  // Fetch all non-draft editions
-  const { data: editionsData, isLoading: editionsLoading } = useQuery({
-    queryKey: queryKeys.editions.all({ pageSize: 50 }),
-    queryFn: async () => {
-      const result = await apiClient.GET('/api/editions', {
-        params: { query: { pageSize: 50 } },
-      });
-      return throwIfError(result);
-    },
-  });
-
-  const editions = editionsData?.data ?? [];
-
-  // Fetch my-submission-status for each edition (parallel)
-  const statusQueries = useQueries({
-    queries: editions.map((edition) => ({
-      queryKey: queryKeys.editions.mySubmissionStatus(edition.id, organizationId ?? ''),
-      queryFn: async () => {
-        if (!organizationId) return null;
-        const result = await apiClient.GET('/api/editions/{id}/my-submission-status', {
-          params: { path: { id: edition.id } },
-          headers: organizationId ? { 'X-Organization-Id': organizationId } : {},
-        });
-        if (result.response.status === 403 || result.response.status === 404) return null;
-        return throwIfError(result);
-      },
-      enabled: !!organizationId,
-    })),
-  });
-
-  // Filter editions where my university has participations
-  const myEditions = editions
-    .map((edition, i) => ({ edition, status: statusQueries[i]?.data }))
-    .filter(({ status }) => status && (status.data?.participations?.length ?? 0) > 0);
-
-  const isLoading = editionsLoading || statusQueries.some((q) => q.isLoading);
+  const { organizationId, currentOrg, myEditions, isLoading } = useDashboardData();
 
   if (!organizationId && !isLoading) {
     return (
