+++ apps/frontend/features/public/auth/register/hooks.ts
@@ -1,0 +1,47 @@
+import { useInvalidateMe } from '@/contexts/AuthContext';
+import { authClient } from '@/lib/auth/client';
+import { useForm } from '@tanstack/react-form';
+import { useState } from 'react';
+import { z } from 'zod';
+
+export function useRegisterForm(onSuccess: () => void) {
+  const invalidateMe = useInvalidateMe();
+  const [error, setError] = useState<string | null>(null);
+
+  const form = useForm({
+    defaultValues: { name: '', email: '', password: '', confirmPassword: '' },
+    onSubmit: async ({ value }) => {
+      setError(null);
+      const result = await authClient.signUp.email({
+        name: value.name,
+        email: value.email,
+        password: value.password,
+      });
+
+      if (result.error) {
+        setError(result.error.message ?? 'アカウント作成に失敗しました');
+        return;
+      }
+
+      await invalidateMe();
+      onSuccess();
+    },
+  });
+
+  return {
+    form,
+    error,
+    validators: {
+      name: z.string().min(1, '名前を入力してください'),
+      email: z.string().email('有効なメールアドレスを入力してください'),
+      password: z.string().min(8, 'パスワードは8文字以上で入力してください'),
+      confirmPassword: ({ value, password }: { value: string; password: string }) => {
+        if (value !== password) {
+          return { message: 'パスワードが一致しません' };
+        }
+
+        return undefined;
+      },
+    },
+  };
+}
