+++ apps/frontend/lib/query/keys.ts
@@ -24,6 +24,8 @@ export const queryKeys = {
       ['editions', id, 'my-submissions', orgId, params] as const,
     mySubmissionStatus: (id: string, orgId: string) =>
       ['editions', id, 'my-submission-status', orgId] as const,
+    submissionMatrixPrefix: (id: string, orgId: string) =>
+      ['editions', id, 'submission-matrix', orgId] as const,
     submissions: (id: string, orgId: string, params: Record<string, unknown>) =>
       ['editions', id, 'submissions', orgId, params] as const,
     submissionMatrix: (id: string, orgId: string, params: Record<string, unknown>) =>
@@ -31,7 +33,10 @@ export const queryKeys = {
   },
 
   participations: {
+    prefix: (id: string, orgId: string) => ['participations', id, orgId] as const,
     detail: (id: string, orgId: string) => ['participations', id, orgId] as const,
+    submissionsPrefix: (id: string, orgId: string) =>
+      ['participations', id, 'submissions', orgId] as const,
     submissions: (id: string, orgId: string, params: Record<string, unknown>) =>
       ['participations', id, 'submissions', orgId, params] as const,
     comments: (id: string, orgId: string, params: Record<string, unknown>) =>
@@ -39,13 +44,18 @@ export const queryKeys = {
   },
 
   submissions: {
+    prefix: (id: string) => ['submissions', id] as const,
+    historyPrefix: (id: string, orgId: string) => ['submissions', id, 'history', orgId] as const,
     history: (id: string, orgId: string, params: Record<string, unknown>) =>
-      ['submissions', id, 'history', orgId, params] as const,
+      [...queryKeys.submissions.historyPrefix(id, orgId), params] as const,
   },
 
   university: {
+    prefix: (orgId: string) => ['university', orgId] as const,
+    membersPrefix: (orgId?: string) =>
+      orgId ? (['university', 'members', orgId] as const) : (['university', 'members'] as const),
     members: (orgId: string, params: Record<string, unknown>) =>
-      ['university', 'members', orgId, params] as const,
+      [...queryKeys.university.membersPrefix(orgId), params] as const,
   },
 
   admin: {
@@ -55,10 +65,14 @@ export const queryKeys = {
     editions: (params: Record<string, unknown>) => ['admin', 'editions', params] as const,
     participations: (editionId: string, params: Record<string, unknown>) =>
       ['admin', 'editions', editionId, 'participations', params] as const,
+    participationsPrefix: (editionId: string) =>
+      ['admin', 'editions', editionId, 'participations'] as const,
     templatesPrefix: (editionId: string) => ['admin', 'editions', editionId, 'templates'] as const,
     templates: (editionId: string, params: Record<string, unknown>) =>
       ['admin', 'editions', editionId, 'templates', params] as const,
+    universitiesPrefix: () => ['admin', 'universities'] as const,
     universities: (params: Record<string, unknown>) => ['admin', 'universities', params] as const,
+    usersPrefix: () => ['admin', 'users'] as const,
     users: (params: Record<string, unknown>) => ['admin', 'users', params] as const,
     userMemberships: (userId: string) => ['admin', 'users', userId, 'memberships'] as const,
   },
